+++ src/media/video/mod.rs
@@ -1,0 +1,20 @@
+// devela::media::video
+//
+//! Video functionality.
+//
+
+crate::items! { // structural access: _mods, _all, _always
+    #[allow(unused)]
+    pub use _mods::*;
+    #[allow(unused)] #[doc(hidden, no_inline)]
+    pub use _always::*;
+
+    mod _mods {
+    }
+    pub(super) mod _all { #[allow(unused)]
+        #[doc(inline)]
+        pub use super::_mods::*;
+    }
+    pub(super) mod _always { #![allow(unused)]
+    }
+}